Move MPU into separate module file
